Muchos cambios. HU1, HU2, HU3, HU5

+++ ms-MCApplication/src/main/java/com/prestabancoms/ms_MCApplication/Repositories/MsMcTypesRepository.java
@@ -4,7 +4,7 @@ import com.prestabancoms.ms_MCApplication.Entities.MsMcTypesEntity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MsMcTypesRepository extends JpaRepository<MsMcTypesEntity, Integer> {
+public interface MsMcTypesRepository extends JpaRepository<MsMcTypesEntity, Long> {
 
     MsMcTypesEntity findByType(String type);
