+++ ms-userRegister/src/main/java/com/prestabancoms/ms_userRegister/Entities/UserEntity.java
@@ -16,10 +16,10 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, nullable = false)
-    private int id;
+    private Long id;
 
     private String name;
     private String password;
-    private int role;
+    private Long role;
 
 }
